fix: remove Bruker 'ser' fallback in search_brucker_binary (#252)
remove the fallback to Bruker 'ser' files in search_brucker_binary.

Only  1D data with 'fid' is currently supported by the spectra client. If a 'ser' file (2D) is found, it is now
skipped and the function returns False. This prevents accidental generation
of 1D spectra from unsupported 2D data.